+++ serialization/tests/TestSerializeCustomHbondForce.cpp
@@ -101,7 +101,7 @@ void testSerialization() {
         ASSERT_EQUAL(b1, b2);
         ASSERT_EQUAL(c1, c2);
         ASSERT_EQUAL(params1.size(), params2.size());
-        for (int j = 0; j < params1.size(); j++)
+        for (int j = 0; j < (int) params1.size(); j++)
             ASSERT_EQUAL(params1[j], params2[j]);
     }
     ASSERT_EQUAL(force.getNumAcceptors(), force2.getNumAcceptors());
@@ -114,7 +114,7 @@ void testSerialization() {
         ASSERT_EQUAL(b1, b2);
         ASSERT_EQUAL(c1, c2);
         ASSERT_EQUAL(params1.size(), params2.size());
-        for (int j = 0; j < params1.size(); j++)
+        for (int j = 0; j < (int) params1.size(); j++)
             ASSERT_EQUAL(params1[j], params2[j]);
     }
     ASSERT_EQUAL(force.getNumExclusions(), force2.getNumExclusions());
@@ -136,7 +136,7 @@ void testSerialization() {
         ASSERT_EQUAL(min1, min2);
         ASSERT_EQUAL(max1, max2);
         ASSERT_EQUAL(val1.size(), val2.size());
-        for (int j = 0; j < val1.size(); j++)
+        for (int j = 0; j < (int) val1.size(); j++)
             ASSERT_EQUAL(val1[j], val2[j]);
     }
 }
